requirements.txt
firebase_admin==6.4.0
Pillow==10.2.0
Requests==2.31.0
streamlit==1.30.0
streamlit_lottie==0.0.5
twilio==9.0.0
pandas==1.2.0
